requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.7
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
